requirements.txt
flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
slack-bolt==1.9.3
openai==1.0.0 